fixed wes' needs
